fix(fe): registerErrorMessageParser 함수 이동
