+++ web/components/Forms/RegisterForm/index.js
@@ -5,8 +5,12 @@ import { Visibility, VisibilityOff } from '@material-ui/icons';
 import { makeStyles } from '@material-ui/core/styles';
 
 import validator from '../../../utils/validator';
-import { registerErrorMessageParser } from '../../../utils/error-parser';
-import { ERROR_DIFFERENT_PASSWORD } from '../../../utils/error-message';
+import {
+  ERROR_DIFFERENT_PASSWORD,
+  ERROR_ID_AND_SUB_EMAIL_DUPLICATION,
+  ERROR_ID_DUPLICATION,
+  ERROR_SUB_EMAIL_DUPLICATION,
+} from '../../../utils/error-message';
 import { SUCCESS_REGISTER } from '../../../utils/success-message';
 import S from './styled';
 import request from '../../../utils/request';
@@ -37,6 +41,23 @@ const initialErrorState = {
   email: '',
   checkPassword: '',
   register: '',
+};
+
+const registerErrorMessageParser = errorMsg => {
+  const errorMsgs = {};
+
+  if (errorMsg === ERROR_ID_AND_SUB_EMAIL_DUPLICATION) {
+    errorMsgs.id = ERROR_ID_DUPLICATION;
+    errorMsgs.email = ERROR_SUB_EMAIL_DUPLICATION;
+  } else if (errorMsg === ERROR_ID_DUPLICATION) {
+    errorMsgs.id = ERROR_ID_DUPLICATION;
+  } else if (errorMsg === ERROR_SUB_EMAIL_DUPLICATION) {
+    errorMsgs.email = ERROR_SUB_EMAIL_DUPLICATION;
+  } else {
+    errorMsgs.register = errorMsg;
+  }
+
+  return errorMsgs;
 };
 
 const RegisterForm = () => {
